+++ lib/controllers/n_switch_account_controller.dart
@@ -1,5 +1,0 @@
-import 'package:get/get.dart';
-
-class NSwitchAccountController extends GetxController {
-  static NSwitchAccountController get to => Get.find();
-}