+++ lib/widgets/n_switch_account.dart
@@ -1,12 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:get/get.dart';
 import 'package:ndk/ndk.dart';
-import 'package:nostr_widgets/controllers/n_switch_account_controller.dart';
 import 'package:nostr_widgets/functions/n_save_accounts_state.dart';
 import 'package:nostr_widgets/l10n/app_localizations.dart';
 import 'package:nostr_widgets/widgets/widgets.dart';
 
-class NSwitchAccount extends StatelessWidget {
+class NSwitchAccount extends StatefulWidget {
   final Ndk ndk;
   final void Function(String pubkey)? onAccountSwitch;
   final void Function(String pubkey)? onAccountRemove;
@@ -25,71 +23,71 @@ class NSwitchAccount extends StatelessWidget {
   });
 
   @override
+  State<NSwitchAccount> createState() => _NSwitchAccountState();
+}
+
+class _NSwitchAccountState extends State<NSwitchAccount> {
+  @override
   Widget build(BuildContext context) {
-    return GetBuilder<NSwitchAccountController>(
-      init: NSwitchAccountController(),
-      builder: (c) {
-        final pubkeys = ndk.accounts.accounts.keys.toList();
-        final loggedPubkey = ndk.accounts.getPublicKey();
+    final pubkeys = widget.ndk.accounts.accounts.keys.toList();
+    final loggedPubkey = widget.ndk.accounts.getPublicKey();
 
-        return Column(
-          crossAxisAlignment: CrossAxisAlignment.stretch,
-          children: [
-            ...pubkeys.map((pubkey) {
-              final isLoggedAccount = loggedPubkey == pubkey;
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.stretch,
+      children: [
+        ...pubkeys.map((pubkey) {
+          final isLoggedAccount = loggedPubkey == pubkey;
 
-              Widget? subtitle;
-              Widget? trailing;
-              void Function()? onTap;
-              if (isLoggedAccount) {
-                trailing = Icon(
-                  Icons.radio_button_checked,
-                  color: Theme.of(context).colorScheme.primary,
-                );
-              } else {
-                subtitle = Align(
-                  alignment: AlignmentGeometry.centerLeft,
-                  child: TextButton(
-                    onPressed: () {
-                      beforeAccountRemove?.call(pubkey);
-                      ndk.accounts.removeAccount(pubkey: pubkey);
-                      nSaveAccountsState(ndk);
-                      c.update();
-                      onAccountRemove?.call(pubkey);
-                    },
-                    child: Text(AppLocalizations.of(context)!.logout),
-                  ),
-                );
-
-                onTap = () {
-                  beforeAccountSwitch?.call(pubkey);
-                  ndk.accounts.switchAccount(pubkey: pubkey);
-                  nSaveAccountsState(ndk);
-                  c.update();
-                  onAccountSwitch?.call(pubkey);
-                };
-              }
-
-              return ListTile(
-                leading: NPicture(ndk: ndk, pubkey: pubkey),
-                title: NName(ndk: ndk, pubkey: pubkey),
-                subtitle: subtitle,
-                trailing: trailing,
-                onTap: onTap,
-              );
-            }),
-            if (onAddAccount != null)
-              Padding(
-                padding: const EdgeInsets.only(top: 16),
-                child: TextButton.icon(
-                  onPressed: onAddAccount,
-                  label: Text(AppLocalizations.of(context)!.addAccount),
-                  icon: Icon(Icons.add_circle_outline),
-                ),
+          Widget? subtitle;
+          Widget? trailing;
+          void Function()? onTap;
+          if (isLoggedAccount) {
+            trailing = Icon(
+              Icons.radio_button_checked,
+              color: Theme.of(context).colorScheme.primary,
+            );
+          } else {
+            subtitle = Align(
+              alignment: AlignmentGeometry.centerLeft,
+              child: TextButton(
+                onPressed: () {
+                  widget.beforeAccountRemove?.call(pubkey);
+                  widget.ndk.accounts.removeAccount(pubkey: pubkey);
+                  nSaveAccountsState(widget.ndk);
+                  setState(() {});
+                  widget.onAccountRemove?.call(pubkey);
+                },
+                child: Text(AppLocalizations.of(context)!.logout),
               ),
-          ],
-        );
-      },
+            );
+
+            onTap = () {
+              widget.beforeAccountSwitch?.call(pubkey);
+              widget.ndk.accounts.switchAccount(pubkey: pubkey);
+              nSaveAccountsState(widget.ndk);
+              setState(() {});
+              widget.onAccountSwitch?.call(pubkey);
+            };
+          }
+
+          return ListTile(
+            leading: NPicture(ndk: widget.ndk, pubkey: pubkey),
+            title: NName(ndk: widget.ndk, pubkey: pubkey),
+            subtitle: subtitle,
+            trailing: trailing,
+            onTap: onTap,
+          );
+        }),
+        if (widget.onAddAccount != null)
+          Padding(
+            padding: const EdgeInsets.only(top: 16),
+            child: TextButton.icon(
+              onPressed: widget.onAddAccount,
+              label: Text(AppLocalizations.of(context)!.addAccount),
+              icon: Icon(Icons.add_circle_outline),
+            ),
+          ),
+      ],
     );
   }
 }