+++ lib/widgets/widgets.dart
@@ -3,3 +3,4 @@ export 'n_picture.dart';
 export 'n_name.dart';
 export 'n_login.dart';
 export 'n_user_profile.dart';
+export 'n_switch_account.dart';